backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
PyMySQL==1.1.1
django-cors-headers==4.4.0
channels==4.1.0
asgiref==3.8.1
python-dotenv==1.0.1
